refactor: split compat route support from server
