+++ packages/app-core/src/api/compat-route-shared.ts
@@ -1,0 +1,113 @@
+import http from "node:http";
+import type { AgentRuntime } from "@elizaos/core";
+import {
+  type ElizaConfig,
+  loadElizaConfig,
+} from "@miladyai/agent/config/config";
+import { isOnboardingConnectionComplete } from "../contracts/onboarding";
+import { resolveExistingOnboardingConnection } from "./provider-switch-config";
+import { sendJsonError as sendJsonErrorResponse } from "./response";
+
+const MAX_BODY_BYTES = 1_048_576;
+
+export interface CompatRuntimeState {
+  current: AgentRuntime | null;
+  pendingAgentName: string | null;
+}
+
+export const DATABASE_UNAVAILABLE_MESSAGE =
+  "Database not available. The agent may not be running or the database adapter is not initialized.";
+
+export function isLoopbackRemoteAddress(
+  remoteAddress: string | null | undefined,
+): boolean {
+  if (!remoteAddress) return false;
+  const normalized = remoteAddress.trim().toLowerCase();
+  return (
+    normalized === "127.0.0.1" ||
+    normalized === "::1" ||
+    normalized === "0:0:0:0:0:0:0:1" ||
+    normalized === "::ffff:127.0.0.1" ||
+    normalized === "::ffff:0:127.0.0.1"
+  );
+}
+
+export async function readCompatJsonBody(
+  req: http.IncomingMessage,
+  res: http.ServerResponse,
+): Promise<Record<string, unknown> | null> {
+  const chunks: Buffer[] = [];
+  let totalBytes = 0;
+
+  try {
+    for await (const chunk of req) {
+      const buf = Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk);
+      totalBytes += buf.length;
+      if (totalBytes > MAX_BODY_BYTES) {
+        req.destroy();
+        sendJsonErrorResponse(res, 413, "Request body too large");
+        return null;
+      }
+      chunks.push(buf);
+    }
+  } catch {
+    sendJsonErrorResponse(res, 400, "Invalid request body");
+    return null;
+  }
+
+  if (chunks.length === 0) {
+    return {};
+  }
+
+  try {
+    const parsed = JSON.parse(
+      Buffer.concat(chunks).toString("utf8"),
+    ) as unknown;
+    if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) {
+      sendJsonErrorResponse(res, 400, "Invalid JSON body");
+      return null;
+    }
+    return parsed as Record<string, unknown>;
+  } catch {
+    sendJsonErrorResponse(res, 400, "Invalid JSON body");
+    return null;
+  }
+}
+
+export function hasCompatPersistedOnboardingState(config: ElizaConfig): boolean {
+  if ((config.meta as Record<string, unknown>)?.onboardingComplete === true) {
+    return true;
+  }
+
+  const existingConnection = resolveExistingOnboardingConnection(
+    config as Record<string, unknown>,
+  );
+  if (isOnboardingConnectionComplete(existingConnection)) {
+    return true;
+  }
+
+  if (Array.isArray(config.agents?.list) && config.agents.list.length > 0) {
+    return true;
+  }
+
+  return Boolean(
+    config.agents?.defaults?.workspace?.trim() ||
+      config.agents?.defaults?.adminEntityId?.trim(),
+  );
+}
+
+export function getConfiguredCompatAgentName(): string | null {
+  const config = loadElizaConfig();
+  const listAgent = config.agents?.list?.[0];
+  const listAgentName =
+    typeof listAgent?.name === "string" ? listAgent.name.trim() : "";
+  if (listAgentName) {
+    return listAgentName;
+  }
+
+  const assistantName =
+    typeof config.ui?.assistant?.name === "string"
+      ? config.ui.assistant.name.trim()
+      : "";
+  return assistantName || null;
+}
