+++ packages/app-core/src/api/database-rows-compat-routes.ts
@@ -1,0 +1,171 @@
+import http from "node:http";
+import type { AgentRuntime } from "@elizaos/core";
+import { ensureCompatApiAuthorized } from "./auth";
+import { DATABASE_UNAVAILABLE_MESSAGE } from "./compat-route-shared";
+import {
+  sendJsonError as sendJsonErrorResponse,
+  sendJson as sendJsonResponse,
+} from "./response";
+import {
+  executeRawSql,
+  quoteIdent,
+  sanitizeIdentifier,
+  sqlLiteral,
+} from "../utils/sql-compat";
+
+export async function handleDatabaseRowsCompatRoute(
+  req: http.IncomingMessage,
+  res: http.ServerResponse,
+  runtime: AgentRuntime | null,
+): Promise<boolean> {
+  const pathname = new URL(req.url ?? "/", "http://localhost").pathname;
+  const match = /^\/api\/database\/tables\/([^/]+)\/rows$/.exec(pathname);
+  if ((req.method ?? "GET").toUpperCase() !== "GET" || !match) {
+    return false;
+  }
+
+  if (!ensureCompatApiAuthorized(req, res)) {
+    return true;
+  }
+
+  if (!runtime) {
+    sendJsonErrorResponse(res, 503, DATABASE_UNAVAILABLE_MESSAGE);
+    return true;
+  }
+
+  const tableName = sanitizeIdentifier(decodeURIComponent(match[1]));
+  const requestUrl = new URL(req.url ?? "/", "http://localhost");
+  const schemaName = sanitizeIdentifier(requestUrl.searchParams.get("schema"));
+
+  if (!tableName) {
+    sendJsonErrorResponse(res, 400, "Invalid table name");
+    return true;
+  }
+
+  let resolvedSchema = schemaName;
+
+  if (!resolvedSchema) {
+    const { rows } = await executeRawSql(
+      runtime,
+      `SELECT table_schema AS schema
+         FROM information_schema.tables
+        WHERE table_name = ${sqlLiteral(tableName)}
+          AND table_schema NOT IN ('pg_catalog', 'information_schema')
+          AND table_type = 'BASE TABLE'
+        ORDER BY CASE WHEN table_schema = 'public' THEN 0 ELSE 1 END,
+                 table_schema`,
+    );
+
+    const schemas = rows
+      .map((row) => row.schema)
+      .filter((value): value is string => typeof value === "string");
+
+    if (schemas.length === 0) {
+      sendJsonErrorResponse(res, 404, `Unknown table "${tableName}"`);
+      return true;
+    }
+
+    if (schemas.length > 1 && !schemas.includes("public")) {
+      sendJsonErrorResponse(
+        res,
+        409,
+        `Table "${tableName}" exists in multiple schemas; specify ?schema=<name>.`,
+      );
+      return true;
+    }
+
+    resolvedSchema = schemas.includes("public") ? "public" : schemas[0];
+  }
+
+  const columnResult = await executeRawSql(
+    runtime,
+    `SELECT column_name
+       FROM information_schema.columns
+      WHERE table_name = ${sqlLiteral(tableName)}
+        AND table_schema = ${sqlLiteral(resolvedSchema)}
+      ORDER BY ordinal_position`,
+  );
+
+  const columns = columnResult.rows
+    .map((row) => row.column_name)
+    .filter((value): value is string => typeof value === "string");
+
+  if (columns.length === 0) {
+    sendJsonErrorResponse(
+      res,
+      404,
+      `No readable columns found for ${resolvedSchema}.${tableName}`,
+    );
+    return true;
+  }
+
+  const limit = Math.max(
+    1,
+    Math.min(
+      500,
+      Number.parseInt(requestUrl.searchParams.get("limit") ?? "", 10) || 50,
+    ),
+  );
+  const offset = Math.max(
+    0,
+    Number.parseInt(requestUrl.searchParams.get("offset") ?? "", 10) || 0,
+  );
+  const sortColumn = sanitizeIdentifier(requestUrl.searchParams.get("sort"));
+  const order =
+    requestUrl.searchParams.get("order") === "desc" ? "DESC" : "ASC";
+  const search = requestUrl.searchParams.get("search")?.trim();
+
+  const filters: string[] = [];
+  if (search) {
+    const likeEscaped = search
+      .replace(/\\/g, "\\\\")
+      .replace(/%/g, "\\%")
+      .replace(/_/g, "\\_");
+    const searchLiteral = sqlLiteral(`%${likeEscaped}%`);
+    filters.push(
+      `(${columns
+        .map(
+          (columnName) =>
+            `CAST(${quoteIdent(columnName)} AS TEXT) ILIKE ${searchLiteral}`,
+        )
+        .join(" OR ")})`,
+    );
+  }
+  const whereClause =
+    filters.length > 0 ? `WHERE ${filters.join(" AND ")}` : "";
+
+  const orderBy =
+    sortColumn && columns.includes(sortColumn)
+      ? `ORDER BY ${quoteIdent(sortColumn)} ${order}`
+      : "";
+  const qualifiedTable = `${quoteIdent(resolvedSchema)}.${quoteIdent(tableName)}`;
+
+  const countResult = await executeRawSql(
+    runtime,
+    `SELECT count(*)::int AS total FROM ${qualifiedTable} ${whereClause}`,
+  );
+  const total =
+    typeof countResult.rows[0]?.total === "number"
+      ? countResult.rows[0].total
+      : Number(countResult.rows[0]?.total ?? 0);
+
+  const rowsResult = await executeRawSql(
+    runtime,
+    `SELECT * FROM ${qualifiedTable}
+      ${whereClause}
+      ${orderBy}
+      LIMIT ${limit}
+     OFFSET ${offset}`,
+  );
+
+  sendJsonResponse(res, 200, {
+    table: tableName,
+    schema: resolvedSchema,
+    rows: rowsResult.rows,
+    columns,
+    total,
+    offset,
+    limit,
+  });
+  return true;
+}
