+++ packages/app-core/src/api/server.workbench-todos-error-handling.test.ts
@@ -4,9 +4,18 @@ describe("handleTaskBackedWorkbenchTodoRoute error handling", () => {
   it("wraps runtime task operations in an operation-specific 500 handler", async () => {
     const fs = await import("node:fs");
     const path = await import("node:path");
-    const serverPath = path.resolve(import.meta.dirname, "server.ts");
-    const source = fs.readFileSync(serverPath, "utf-8");
+    const serverSource = fs.readFileSync(
+      path.resolve(import.meta.dirname, "server.ts"),
+      "utf-8",
+    );
+    expect(serverSource).toContain(
+      "handleWorkbenchCompatRoutes(req, res, state)",
+    );
 
+    const source = fs.readFileSync(
+      path.resolve(import.meta.dirname, "workbench-compat-routes.ts"),
+      "utf-8",
+    );
     const handlerIdx = source.indexOf(
       "async function handleTaskBackedWorkbenchTodoRoute",
     );
